pyproject.toml
[project]
name = "mcp-web-gateway"
version = "0.0.0dev"
description = "Connecting MCP Agents to 100s of Web Services "
readme = "README.md"
requires-python = ">=3.12"
license =  {text = "MIT"}
authors = [
    { name = "Tapan Chugh" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastmcp @ git+https://github.com/jlowin/fastmcp.git",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.30.0",
    "fastapi>=0.100.0",
]
agent = [
    "fast-agent-mcp>=0.2.46",
]

[project.urls]
"Homepage" = "https://github.com/chughtapan/mcp-web-gateway"
"Bug Reports" = "https://github.com/chughtapan/mcp-web-gateway/issues"
"Source" = "https://github.com/chughtapan/mcp-web-gateway"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_web_gateway"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
addopts = "--ignore=tests/integration"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.12"
mypy_path = "src"
exclude = "^tests"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_subclassing_any = true
disallow_untyped_calls = true

[tool.flake8]
max-line-length = 80
extend-select = "B950"
extend-ignore = "E203,E501,E701,W291"

